pyproject.toml
[project]
name = "desktop-agent"
version = "0.1.0"
description = "PC常駐型AIアシスタント - タスク管理、スケジュール管理、身の回りのサポートを提供するインテリジェントなデスクトップエージェント"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "toml>=0.10.2",
    "httpx>=0.25.0",
    "click>=8.1.0",
    "rich>=13.0.0",

    "docker>=6.1.0",
    "pydantic>=2.0.0",
    "asyncio-mqtt>=0.11.0",
    
    # MCP関連
    "mcp[cli]>=1.0.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.scripts]
desktop-agent = "cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "cli", "api"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
